small fixes

+++ esp32_utils.cpp
@@ -19,9 +19,9 @@ esp_reset_reason_t ESP32Utils::rebootReason() {
     return esp_reset_reason();
 }
 
-uint32_t ESP32Utils::millis() {
+uint64_t ESP32Utils::millis() {
     static const uint64_t msFactor {1000ULL};
-    return static_cast<uint32_t>(esp_timer_get_time()/msFactor);
+    return static_cast<uint64_t>(esp_timer_get_time()/msFactor);
 }
 
 uint64_t ESP32Utils::micros() {
